Move launch parameter sanitation to separate files
